Add files via upload
Considerations (taken from the exemplar of what the outcome this activity is supposed to help me learn/understand)

What are some key takeaways that you learned from this lab? 
* A lot of machine learning workflows are about cleaning, encoding, and scaling data. * 
The approach you use to plot or graph your data may depend on the type of variable you are evaluating. * Training a logistic regression model on a single independent variable can produce a relatively good model (80.2 percent accuracy).

What findings would you share with others? 
* Logistic regression accurately predicted satisfaction 80.2 percent of the time.
* The confusion matrix is useful, as it displays a similar amount of true positives and true negatives.
What would you recommend to stakeholders? 
* Customers who rated in-flight entertainment highly were more likely to be satisfied. Improving in-flight entertainment should lead to better customer satisfaction. 
* The model is 80.2 percent accurate. This is an improvement over the dataset’s customer satisfaction rate of 54.7 percent. 
* The success of the model suggests that the airline should invest more in model developement to examine if adding more independent variables leads to better results. 
Building this model could not only be useful in predicting whether or not a customer would be satisfied but also lead to a better understanding of what independent variables lead to happier customers.